Make attrs floats

+++ gunslinger/g1/AiPlayer.java
@@ -28,7 +28,7 @@ public class AiPlayer {
 
   public boolean[] enemies, friends;
 
-  public int[] attrs;
+  public float[] attrs;
   public float[] coeffs;
 
   public AiPlayer(int n, int e, int f, float[] cs){
@@ -38,7 +38,7 @@ public class AiPlayer {
       max_e = e;
       max_f = f;
 
-      attrs  = new int[]{ 0, 0, 0, 0, 0, 0, 0};
+      attrs  = new float[]{ 0, 0, 0, 0, 0, 0, 0};
       coeffs = cs;
 
       enemies = new boolean[n];
@@ -53,7 +53,7 @@ public class AiPlayer {
       max_e = e;
       max_f = f;
 
-      attrs  = new int[]{ 0, 0, 0, 0, 0, 0, 0, 0};
+      attrs  = new float[]{ 0, 0, 0, 0, 0, 0, 0, 0};
 
       coeffs = new float[]{-10f, 5.6f, 1.2f, 1.2f, 5.6f, -2.8f, 10.0f, 0.5f};
